Fix sizing and drawing problems of the info window
The info window now only adds a scrolled window if it is really needed (depending on the added widget). This solves the problem that the window was not shrinking correctly to fit for instance a label (useful for calltip like info windows). It also fixes calculating the correct window size (although it might need some more work) and correct referencing of the added widget.