Rework of proposal populating and modelling.
This is a big rework of how proposals are populated and how
the whole workflow of completion works. There is now a special
'context' object which glues together the completion and the providers.
This allows proposals to be added asynchronous and better handling of
provider matching and passing context information for proposal
populating.