Special case the bracket matching tag when printing
Back when we created the compositor, we didn't treat bracket match tags
in a special way, because we wanted a more generic way to mark certain
tags for printing or not. But given that we didn't add that api and that
in all this time bracket match was the only problem reported, go for the
special case solution.