Stop matching bracket search when going out of context class
This stops the matching bracket search if the set of context
classes that apply to the matching end iter is _smaller_ then the
set of classes that apply to the start matching iter.

In effect this means that you stop search if you go outside of the
scope (in terms of context classes) that apply to the starting
bracket, and thus will only match brackets in the same string
or comment (this does not affect outside-of string/comment
brackets).

Note: This works well with the current classes and how they are
defined but might need some refinement if we are adding
more classes.