Install Glade catalog, fixing GNOME bug #527137
Add a configure option --enable-glade-catalog. If enabled, check for the
presence of a suitable version of Glade and install the catalog file to
the correct location, as determined by the catalogdir variable in the
Glade pkg-config file. The pkg-config check was taken from mm-common,
and originally written by Daniel Elstner.