Always consume left click gutter button press event
This is technically not really correct, but nobody else is using the
event anyway. The more fundamental issue is that the cell-activated
signal does not allow a return value indicating the event should be
consumed. This caused the line gutter clicking things (select line,
extend selection) to fail because the buffer got the event and removed
the selection.