Fixed scrolling after an undo or redo
Scrolling after an undo or redo was broken now we do the following:

If the insert mark is hidden vertically then scroll to it
and align it to the middle of the X and Y axis.

Otherwise, if the insert mark is hidden horizontally then scroll to it
and align it to the middle of the X axis. Because we cannot prevent it from
also aligning to the Y axis we revert the vertical position of the view.