Update of M4 syntax highlighting
Brought the M4 language definition up-to-date with all current
Autoconf-and-friends macros. Added an exhaustive list of macros with
information on how to highlight their arguments (i.e. whether they
contain shell code or not.) This functionality was already present in
the file, but used only for a few macros. Commented the file.