Changed "bracket-matched" signal.
Now this signal's 'state' parameter is a value instead of a pointer,
so it is consistent with what documentation says and what is passed
to g_signal_new(). Also, this signal is not emitted if previous and
current char are nonbrackets. Added a function checking this to
C and python test-widget.