Rip out custom CompletionInfo sizing.
It is totally broken with gtk3 new sizing logic so we would need to redo
it from scratch. Beside it had some dubious code like the one to decide
if a scrolledwindow was needed. Beside the completion info is a gtkwindow,
so worse case apps can do the sizing themselves.