win32: fix sed portability issue
When editing files in-place (sed -i) on windows, sed cannot
open the input file for read and for write at the same time.
We can fix this by not editing in-place but feed the input
file into sed's stdin and write stdout to a temporary file.
When done rename the temporary file on top of the input file.

Doing so achieves the same result and fixes the win32 sed
portability issue.