test-completion: a fixed and a random provider
For testing the bug #629055 [1], the proposals need to be modified
when the user types something and the completion window doesn't
disappear.

For the random provider, the proposals are choosen at random from a
fixed set of 10 proposals. This is done on each populate.

With 'show-headers' set to true, there is no problem. If it is set to
false, there are some critical messages, and sometimes a segfault.

The random provider have a lower priority than the fixed one, because
the display is more "stable" with the fixed one first.

[1] https://bugzilla.gnome.org/show_bug.cgi?id=629055