CompletionModel: fix unmarked nodes removal
Take into account the fact that the nodes can be filtered.

Also, after removing the unmarked nodes, if there is no remaining nodes,
the header was removed only if 'show-headers' is true. But if
'show-headers' is false, the node is still there (but filtered) and
needs to be removed too.

This finally fixes bug #629055 "GtkSourceCompletion show-headers
property: Critical messages and crash":

https://bugzilla.gnome.org/show_bug.cgi?id=629055