Completion doc: "provider-page-size" is not yet used
The "provider-page-size" property is actually not used. It could be used with
Ctrl+PageDown and Ctrl+PageUp. But would it be very useful? Generally
there are not a lot of providers AFAIK. So the property could simply be
removed or deprecated.

Anyway I think it's better to document the fact that it is not yet used.