Completion: fix "Alt" keyboard shortcuts
Alt+1, Alt+2, ... are used to activate one of the first proposals. These
shortcuts didn't work with some keyboard layouts because the Shift key
must be pressed to access the digit.

The variable 'mod' in this case is not exactly equal to GDK_MOD1_MASK
(the Alt key), since the bit for Shift is also enabled.