Better return type for gtk_source_completion_get_view()
The return type of gtk_source_completion_get_view() is
"GtkSourceView *". But due to a cycle in the .h includes, the type was
"struct _GtkSourceView *".

The header includes cycle: gtksourceview.h needs the GtkSourceCompletion
type. But gtksourcecompletion.h needs the GtkSourceView type.

The solution is to put the typedef of GtkSourceCompletion in
gtksourcecompletion-private.h and include this header both in
gtksourceview.h and gtksourcecompletion.h.

Redifining two times the same type with typedef is not allowed in C.