CompletionInfo: always a natural size
The previous implementation had the public function:

set_sizing (int max_width, int max_height, bool shrink_width, bool
shrink_height);

It has been removed during the port to GTK+ 3. See commit
e195094a2a89fe25996f081b584c42675b89216f.

However, the sizing of a GtkWindow should be done with GTK+ functions,
not by adding API to GtkSourceView.

Previously, the info window had by default a fixed size of 300x200, with
a scrolled window. It was difficult to change this for having a natural
size. Now the natural size is the default, and having a fixed size is
not really difficult (it will be documented in a later commit).

For more details, read:
https://bugzilla.gnome.org/show_bug.cgi?id=679280