GTK-Doc: extra directories in FIXXREF_OPTIONS
The links to external GNOME symbols works now. It is how it is done in
GTK+.

To avoid the warnings, the HTML documentation of GLib, GTK+ and Pango
must be available, obviously. A quick way to add them, without
rebuilding everything, is to install the doc packages from the
distribution, and add symbolic links in the jhbuild prefix.