Use G_GNUC_INTERNAL to avoid exporting private symbols
With the relinfo.pl script from Ulrich Drepper, here are the results.

Before:
128 relocations, 105 relative (82%), 946 PLT entries, 247 for local syms (26%)

After:
125 relocations, 103 relative (82%), 880 PLT entries, 181 for local syms (20%)

Private symbols prefixed with an underscore (or without the gtk_source_
prefix) were not exported. However G_GNUC_INTERNAL is beneficial for
them too because the compiler is now aware that those symbols are not
exported, and therefore it can better optimize the code.

Another problem remains: the big number of PLT entries for local
symbols. Normally it should be 0 I think (it is 0 for the glib, gio and
gtk+, for example, so it should be possible for GSV too).