CompletionModel: new implementation
The previous implementation supported several populations of proposals.
Between two populations, the model kept the common proposals, to avoid
emitting useless "row-deleted" and "row-inserted" signals.

But there is a simpler solution. During a population, we remove the
model from the GtkTreeView. So the "row-inserted" and "row-deleted"
signals are not needed. Once the population is finished, we set the
model to the GtkTreeView, so the rows are correctly displayed.

With the new implementation, a model is used for only one population.
When a new population occurs, the old model is destroyed, and we create
a new one. It simplifies a lot the code.

And it's also more efficient, since we avoid completely all the
"row-inserted" and "row-deleted" signals during a population.

I've tested with five providers containing each 10000 proposals. With
the previous implementation, the completion window takes maybe 30
seconds to appear, while with the new implementation, it appears in
approximately one second!

There is still the problem when we change the visible providers. In this
case the "row-deleted" and "row-inserted" signals are emitted. But it
would not be difficult to use the same trick as for the populations.