Completion: cleanup cancellation and updating of the current context
Remove the context parameter from cancel_completion() and change
it so that it only unrefs and sets the active context to NULL.

update_completion() now always expect that it owns a
ref on the passed in context. This means that cancel_completion() always can unref
the previous active context even if the new context passed in is the same.

This should fix a potential problem where the active context wasn't updated properly
leading to the context being leaked and the active context being NULL even though
it shouldn't.

https://bugzilla.gnome.org/show_bug.cgi?id=691668