Completion: use a new model for each population
At the beginning of a population, we remove the model from the TreeView,
and we destroy it. A new model is created, which is set to the TreeView
only when the population is finished.

When the population is finished, we must also show the completion
window, which is normally done in on_row_inserted_cb(), but the
"row-inserted" signals are not emitted.