Completion: separate CompletionPrivate fields into sections
The fields are not renamed, they are just moved, and some comments are
added to explain them.

For the widgets, some IDs in the UI file are renamed, to have the same
name as the struct field, or to have a better name (the struct fields
will be renamed later in this case).