UndoManagerDefault: don't use the INVALID action
An INVALID action was used to distinguish from a NULL action in the
'modified_action' struct field.

But using NULL instead of INVALID works fine. The only place where an
INVALID value would make a different result is in
modified_changed_handler():

if (manager->priv->modified_action != NULL)
{
        g_message ("%s: oops", G_STRLOC);
        return;
}

If the rest of the code is correct, it doesn't change anything.

https://bugzilla.gnome.org/show_bug.cgi?id=694613