Use GLIB_VERSION_MIN_REQUIRED, GLIB_VERSION_MAX_ALLOWED, etc
It will report warnings when the required version of GLib or GTK+ is not
correct.

For example if we use a new GTK+ function without bumping the dependency
version, a warning will be printed.

It assumes that GLIB_CHECK_VERSION() is not used, so the minimum
required version is the same as the maximum allowed one.