Show mnemonics in the main completion window
Before, the mnemonic for the Details button was always hidden, even if
we pressed Alt, because the completion window doesn't have the focus.

It would be possible to show the mnemonic only at appropriate times, but
it would be too difficult for such a small detail (I know, Every Details
Matters, but simple code matters too :).

So it's better to always have the mnemonic visible, than never.