Completion: remove interactive_providers struct field
When a provider is added to Completion, it is added in the list of all
providers, in the 'providers' struct field. It was also added to the list
of interactive providers if the provider supports the interactive
activation.

The problem is that if the provider wants to change its mind and wants
to add or drop support of interactive activation, _after_ it has been
added to the Completion, this didn't work as expected.

A workaround was to remove and re-add the provider to the Completion,
but it is really a hack.

So it's better, when starting an interactive completion, to always check
all the providers.

Example of use:
https://git.gnome.org/browse/latexila/commit/?id=69449d1b00e1d980a36ec64e7024bc2a6d4bf31a