Completion: fix the update of the window position
The window position was updated too many times. In some situations, we
could see the intermediate positions, if these intermediate positions
were not the same as the final position. So when typing characters at a
certain speed, the completion window was shaking. Although it's quite
funny, it's better if it is fixed :)