Completion: hide bottom bar in some cases
If there is only one provider, the label "All" in the bottom bar (on the
right) is useless.

If there is no extra information, the Details button is useless.

When both widgets are useless, we hide the bottom bar.

The main completion window has always the same size, so when the bottom
bar is hidden, the GtkTreeView is bigger.