Fix build with -Wl,--as-needed in the LDFLAGS
The *.la files listed in the LDADD variable were in the wrong order.

If libgtksourceview-3.0.la is before libgtksourcecompletionwords.la,
the linker doesn't add the DT_NEEDED tag for libgtksourceview-3.0.la
because at that point in the linkage, libgtksourceview-3.0.la is
useless.

On the other hand, if libgtksourcecompletionwords.la is listed before
libgtksourceview-3.0.la, the DT_NEEDED tag is added to the latter
because the former needs a symbol from the latter.

Thanks to Dominique Leuenberger for the bug report.

https://bugzilla.gnome.org/show_bug.cgi?id=701113