CompletionWords buffer: use GtkTextRegion
For scanning the words (for the initial scan but also on text insertion
and text deletion), the "scan region" is now a GtkTextRegion. It
simplifies the code for maintaining the GtkTextMarks.

For modifications on the buffer, there is now a callback before the
modification, to invalidate the region (remove the words that must be
removed), and a callback after the modification, to add the region to the
scan region. There were some problems when adding the region to the scan
region before the modification (text insertion or deletion).