Add API to enable/disable search highlighting
And add a button in test-search-ui to test the new setting.

Disabling the search highlighting can be useful for using the API in a
different context than the standard search and replace. For example the
gedit spell checking plugin can use the replace_all() function.

It is also useful to be able to disable/enable the search highlighting
without losing the text to search.