New search and replace API, with SearchContext and SearchSettings
The API is more flexible. Before, only one search was possible at a
time: there was a 1:1 relationship between the buffer and the private
Search object. Now it is a 1:N relationship between the buffer and the
SearchContext: it is possible to run several searches at the same time.
And also a 1:N relationship between the SearchContext and the
SearchSettings: it is possible to share search settings between several
buffers.

In summary, it is an indirect many-to-many relationship between the
buffer and the search settings.

       0-N       1-1               1-1       0-N
Buffer ------------- SearchContext ------------- SearchSettings