SearchContext: init found_tag in set_buffer()
With the previous API, the buffer could not be completely constructed
when the Search object was created, so the solution was to call
init_found_tag() a bit everywhere, if the found_tag was not already
initialized.

Now when the buffer is set, we are sure that it is constructed, so we
can directly init the found_tag without problems with the tag table.