Add gtk_source_buffer_disable_search_highlighting()
The buffer must have a list of search contexts. It is done with
_gtk_source_buffer_add_search_context(), and the buffer holds a weak
reference to the context. It must not be the other way around, the
context must have a strong reference to the buffer, because the context
needs the buffer. On the other hand, the buffer can work without search
contexts.