SourceStyle: Use the expected typedef for GtkSourceStyleClass.
typedeffing GtkSourceStyleClass to GObjectClass was a shortcut
that was acceptable when hidden in the .c file, but putting it in
the header conflicts with the expectation that it will be declared
like other such *Class structs. This fixes that, so gtksourceviewmm
can build again, because gtksourceviewmm has to duplicate the
(expected) declaration.