Update the full lang-spec file to match the tutorial text
The full language definition file attached at the end of the tutorial is
updated to match the snippets in the tutorial text.  Changes include:
- Use <metadata> for mimetypes and filename extensions;
- Delete "text/x-chdr" mimetype;
- Add "comment" style tag;
- Use context ID "escape" instead of "escaped-character".
- Use "string" style instead of "char" style for "char" context.
- Move "if0-comment" context section before "preprocessor" context so
  that it has higher priority to be matched;
- Escape character "&" in "&lt;" and "&gt;";
- Fix typo "inlcuded" -> "included";

The coding style is kept as close as possible to the original, although
the original had inconsistent style to begin with (e.g. 2-space indent
mixed with 4-space indent).

Also see:
https://bugzilla.gnome.org/show_bug.cgi?id=707372