Style schemes: more visible right margin
The alpha for the foreground (the line) is 16%. The alpha for the
background (the area on the right of the line) is 6%. So for a light
style, dark colors must be set for the right margin. And for a dark
style, light colors must be set.

It's better to not adjust the alpha channel in the code. Maybe
third-party style schemes rely on these values.