Remove second parameter of gtk_text_region_destroy()
It was a boolean to know if the marks owned by the GtkTextRegion must be
removed from the buffer. In every case we want to remove the marks,
except when the buffer is already finalized, in which case FALSE is
passed as the second parameter.

But now GtkTextRegion has a weak ref to the buffer, so the destroy
function knows when the buffer is finalized.