Completion: do not show "All" when only one provider
In the bottom bar of the completion window, at the right of the Details
button, there is an icon and a label to show on which provider we are.
When there is only one provider, all the proposals are shown. Displaying
the "All" label in this case is useless, and users don't know what it
means. So it's better to display nothing in this case.