SearchContext: do not emit the highlight-updated signal
When the search was implemented in gedit, there was the
search-highlight-updated signal, similar to the highlight-updated signal
from GtkSourceBuffer. So when moving the search implementation in
GtkSourceView, I reused the highlight-updated signal.

I think the signal is no longer required for the search. I added a
comment for the highlight-updated signal for explaining when this signal
is useful (and currently used in the spell gedit plugin).

Moreover, the search didn't send the signal at the right time. When
g_signal_emit_by_name() was called, the start and end variables could be
the entire buffer, resulting in performances problems when the spell
gedit plugin was enabled with the Highlight Misspelled Words option
(because the plugin connects to the signal, and rescan the specified
region).

https://bugzilla.gnome.org/show_bug.cgi?id=722535