Revert "Prevent construction of completion after dispose has run"
This reverts commit 89382601e772c0504ae7e5f709399f4dd3b0d291.

It was used in get_completion() to prevent the creation of
GtkSourceCompletion if dispose() has already run. But if
get_completion() is called in such a (rare) situation, it means that
GtkSourceView is still alive, and that dispose() will anyway be run a
second time, where g_clear_object() will clear the completion object a
second time if needed.

There was before a reference cycle between the view and completion, now
it is not the case anymore. This was maybe the root of the problem.

I've tested this commit with the following program, and everything is
fine, the completion is created and freed two times.

gint
main (gint argc, gchar *argv[])
{
        GtkSourceView *view;

        gtk_init (&argc, &argv);

        view = GTK_SOURCE_VIEW (gtk_source_view_new ());
        g_object_ref_sink (view);

        gtk_source_view_get_completion (view);

        g_message ("main: run dispose on view");
        g_object_run_dispose (G_OBJECT (view));

        gtk_source_view_get_completion (view);
        g_message ("main: unref view");
        g_object_unref (view);

        return 0;
}