Completion: weak reference to the view
There was a reference cycle between the view and completion. Now only
the view has a strong reference to the completion. It is more logical
that the view owns the completion, not vice-versa.

This wasn't really a problem as far as the view was added to a
container. Because when the container is destroyed, it calls
gtk_widget_destroy() on its children, which calls g_object_run_dispose()
which breaks the reference cycle.

Anyway, it's better to not rely on gtk_widget_destroy(), and thus avoid
reference cycles in C code.

The drawback is that there are now checks to see if view == NULL, before
using view.