View: do not connect to the highlight-updated signal
This code was added in the big commit "towards gedit integration" back
in 2003, without a comment explaining why this code was required.

I think it is no longer required. If a GtkTextTag is added to the
buffer, the view will normally be redrawn. And if a GtkTextTag is
already present in the buffer, but one of its property changes, the view
will normally also be redrawn.

It was perhaps used for updating the gutter, the right margin, etc. But
it seems to work fine without calling gtk_widget_queue_draw_area().

If something breaks, we can revert this commit, and add a comment
explaining why it is useful.