CompletionContext: 'user requested' default activation
A None activation currently doesn't match any provider. To match at
least one provider, you have to change the activation to 'interactive'
or 'user requested'. The 'user requested' activation is more useful
outside of gsv (for example for another keybinding than Ctrl+space), so
it is now the default value. It doesn't really break the API.

With the None activation, there is also a possible ambiguity: does it
mean "I don't care about activation", or does it mean "The completion is
not activated"? Both make sense. With the 'user requested' activation,
there is no ambiguity.

And if the user of the API don't care about the activation, it will work
as expected now, since the provider returns both activations by default.