Fix testdark.xml style scheme
testdark.xml is not installed, it is just a test. But some unit tests
call gtk_source_style_scheme_manager_set_search_path() with the
data/styles/ directory from the sources, so the unit tests can run when
gtksourceview is not installed. So testdark.xml in this case is read,
and there was a problem with the colors not recognized (the colors are
not inherited from the parent scheme).

https://bugzilla.gnome.org/show_bug.cgi?id=731039