words-completion: 'end' parameter for scan_line()
Currently the words are scanned line by line. The goal is to be able to
scan by words on text insertion and deletion. So a first step is to have
an 'end' iter for scan_line(), so it stops at @end.

Add also 'const' for some GtkTextIter parameters that are not modified.

This commit normally fixes:
https://bugzilla.gnome.org/show_bug.cgi?id=725729