words-completion: work by words instead of lines
For example, if I insert a character in this line (at the cursor
position |):
word1 word2 | word3

Before this commit: before the text insertion, word1 word2 and word3 are
removed from the library. If these words are not present elsewhere in
the buffer, they are not available for completion when I insert the
text.

After this commit: instead of working line by line, the region to scan
(either for removing or adding words) is adjusted to word boundaries,
not line boundaries. So word1, word2 and word3 are available for
completion.