CompletionContainer sizing fixes
The previous implementation subclassed GtkBin and added/removed the
scrolled window during the size requisition, which is probably one of
the problem. Now the CompletionContainer is a subclass of
GtkScrolledWindow, so the scrolled window is always there, with an
automatic policy.

Another source of bug is that the code to know whether a
scrollbar is needed was buggy and not consistent between
get_preferred_width() and get_preferred_height(). It should now be
fixed.

https://bugzilla.gnome.org/show_bug.cgi?id=720823