gutter: always invalidate the whole GdkWindow
Call gdk_window_set_invalidate_handler() in update_gutter_size(),
because the GdkWindow is created only when its size is greater than 0.
update_gutter_size() is already called when the widget is realized, so
it ensures that the invalidate handler is set at least once.

The performances should not be a problem, because the gutter was already
redrawn completely on each mouse motion event in the gutter (not when
moving the mouse in the text view).

https://bugzilla.gnome.org/show_bug.cgi?id=732418