buffer: fix default value of max-undo-levels
The documentation said that 1000 was the default value, but
max-undo-levels is not a construct property, and the 1000 value was not
set in init(). So in reality the default value was 0. And the default
undo manager has a bug, 0 is the same as -1 (i.e. unlimited).

So by default the undo/redo was unlimited. Now the doc and the code
reflect that.