php.lang: relicense from GPL to LGPL
All the authors accept to relicense to LGPL, except Rowan Lewis who
doesn't reply to my mails. Rowan Lewis's contribution is in the
commit 837bc65f31435571942aa9d65a4e5c5ead7c8db6, which adds 6 keywords
to php.lang. For such a small change the copyright doesn't really hold.
It is explained for example in the GNU Coding Standards:

"You don't need papers for changes of a few lines here or there, since
they are not significant for copyright purposes."

http://www.gnu.org/prep/standards/standards.html

So I remove Rowan Lewis's copyright, and relicense php.lang to LGPL.

https://bugzilla.gnome.org/show_bug.cgi?id=159134