Better default candidate encodings, with translations
The translations are available in gedit. It is something difficult to
translate, so it's better to move that in GtkSourceView.

There is a new public function in GtkSourceEncoding to get the default
candidates. The FileLoader uses by default that list, and adds the
GtkSourceFile's encoding if it has been set by a previous file loading
or saving operation. The file's encoding is added to the beginning of
the list.