sh.lang: add globs for bashrc, .profile and .bash_profile
~/.bashrc, /etc/bashrc and /etc/bash.bashrc are covered by "*bashrc".
The likelihood that a simple text file ending with "bashrc" is low, so
"*bashrc" should be fine.

However for ~/.profile and ~/.bash_profile, with the glob "*profile"
there are more likelihood that it'll wrongly match simple text files. So
the file names ".profile" and ".bash_profile" are matched explicitly.
The drawback is that /etc/profile is not covered.

https://bugzilla.gnome.org/show_bug.cgi?id=733624