UndoManager: new implementation based on GQueues
See also the previous commit message.

Visible changes with the previous implementation:
- Word merging: before, the spaces _after_ a word were merged with the
word, now it is the spaces _before_ the word that are merged. See the
comment in action_insert_merge() for the reason.

- A group containing several actions will now never be merged with the
previous group. Before, the merging was done action by action.

- An insertion or deletion outside a begin/end user action is now
inserted into a new group for each action. With the previous
implementation, those actions were grouped into one undo/redo step.

- Different behavior for restoring the cursor position. The selection is
no longer restored. The cursor position is placed at the end of the
first action in the group. If the group contains only one action, it
simply goes to the end of the insertion or deletion. For a search and
replace, it always goes to the end of the first replaced occurrence.
If restoring the selection is desirable, the code contains #if 0
sections for storing the selection only if the selection is related to
the action. That is, for a deletion the selection bounds should have the
same offsets as the deleted text, otherwise the selection is not stored.