docs: improve a little the GtkSourceBuffer doc
Reorganize the order of the functions. The class description describes
the undo/redo and then the context classes, so apply the same order.
Syntax highlighting is more important and is thus at the beginning.
Marks and other functions are at the end.

The generated html doesn't display the subsections. GTK-Doc could
separate the different sections into blocks, and display the titles.