Test UndoManager performances
And the results are....

Old implementation based on a GPtrArray:
Undo/Redo 10000 actions: 0.486454 seconds.
Undo/Redo 100000 actions: 4.822937 seconds.

New implementation based on GQueues (see next commits):
Undo/Redo 10000 actions: 0.494246 seconds.
Undo/Redo 100000 actions: 4.903951 seconds.

So the difference is small, it's linear with the number of undo's, which
means that an undo or redo is a O(1) operation. It was not the case with
the older implementation based on GList, because GList was not used
correctly (g_list_nth() was called on each undo or redo, because the
location in the history was not stored with a GList node, but with a
gint index).