build: dependencies more at the top of configure.ac
So that the top of configure.ac contains all the important stuff that is
often changed.

The Libtool versioning is changed only once every six months, but it's
important to not forget it, so it's still at the top too. (and the algo
to change the libtool versioning comes actually from the GNU
documentation, which was copied in the Autotools book).