docs: clarification of gtk_source_gutter_renderer_draw()
"without padding or margin" can be interpreted differently: does it mean
the padding is not taken into account so it actually _includes_ the
padding, or does it mean that the padding is outside the
background_area?

It turns out that "without padding" meant that the padding is included
in the area delimited by background_area.