Completion: don't use gtk_style_context_get_background_color()
The function is deprecated.

With Adwaita the background color was anyway white, so it had no
effects. With some early GTK+ 3 versions the background color was a
little different, see for example this screenshot:
http://blogs.gnome.org/swilmet/files/2013/03/completion_calltip_before.png
(see the background color of the header, "LaTeX")

The best would be to write some CSS that we attach to the GtkTreeView,
so we apply the exact style that we want for the headers and
accelerators, but I think CSS is not really supported by GtkTreeView.

Using the GtkCellRenderer:sensitive property almost works, but a detail
is different: the accelerator in the selected row is difficult to see,
since the digit is still gray. For unselected rows the accelerators must
be gray, but for the selected row the accelerator must have a white
foreground color so it's clearly visible with the blue background color
(with Adwaita).