Make properties descriptions untranslatable
It's useless to translate the API. Properties are part of the API, it is
not seen by normal users, only by developers. GTK-Doc comments are
anyway not translated, so it's a bit strange to translate properties
while all the rest of the developer documentation is in English.

Moreover, translating properties is probably a difficult exercise. Some
comments for translators were needed. So it doesn't really worth the
effort.